requirements.txt
appdirs==1.4.4
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
idna==2.10
importlib-metadata==4.6.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
pkg-resources==0.0.0
regex==2021.4.4
requests==2.25.1
rpi==0.0.1
RPi.GPIO==0.7.0
spidev==3.5
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.4.1
